% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometric_space_qualitative.R
\name{ecometric_space_qual}
\alias{ecometric_space_qual}
\title{Plot Ecometric Space for Qualitative Traits with Optional Fossil Overlay}
\usage{
ecometric_space_qual(
  model_out,
  category_labels = NULL,
  palette = NULL,
  fossil_data = NULL,
  fossil_color = "black",
  modern_color = "#bc4749"
)
}
\arguments{
\item{model_out}{Output from \code{ecometric_model_qualitative()}, containing environmental estimates in trait space.}

\item{category_labels}{Optional named vector for category labels (for prettier plots).}

\item{palette}{Optional color vector for categories (must match number of categories).}

\item{fossil_data}{Optional. Fossil data from \code{reconstruct_env_qualitative()}, including \code{fossil_mbc} and \code{fossil_sdc}.}

\item{fossil_color}{Outline color for fossil data bins (default = "#c44536").}

\item{modern_color}{Outline color for modern data bins (default: \code{"#bc4749"}).}
}
\value{
A list containing:
\item{ecometric_space_plot}{ggplot showing the predicted category across trait space.}
\item{probability_maps}{List of ggplots showing probability surfaces for each category.}
}
\description{
Visualizes the predicted ecometric space (dominant category) and probability maps
for each category based on the output from \code{ecometric_model_qualitative()}.
}
