% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fossils}
\alias{fossils}
\title{Fossil trait data for projection onto ecometric space}
\format{
A data frame with the following columns:
\describe{
\item{Site}{Unique identifier for the fossil sample}
\item{fossil_summ_trait_1}{Estimated mean of relative blade length for the fossil site}
\item{fossil_summ_trait_2}{Estimated sd of relative blade length for the fossil site}
\item{Long}{Longitude coordinate (decimal degrees)}
\item{Lat}{Latitude coordinate (decimal degrees)}
}
}
\source{
Siciliano-Martina et al. (2024). \emph{Ecology and Evolution}, 14(10), e70214.
}
\usage{
fossils
}
\description{
A dataset of fossil sites with estimated trait distribution and geographic coordinates,
used to project extinct communities onto modern ecometric space.
}
\keyword{datasets}
