% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_bins.R
\name{optimal_bins}
\alias{optimal_bins}
\title{Determine optimal number of bins using Scott's Rule}
\usage{
optimal_bins(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Integer representing the optimal number of bins.
}
\description{
Calculates the optimal number of bins for a numeric vector
based on Scott's rule. For more details see Scott (1979) \url{doi:10.1093/biomet/66.3.605}
}
\examples{
# Simple example
# Example with normally distributed data
optimal_bins(rnorm(100))
}
