% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_grid.R
\name{optimal_bins_scott}
\alias{optimal_bins_scott}
\title{Determine optimal number of bins using Scott's Rule}
\usage{
optimal_bins_scott(x)
}
\arguments{
\item{x}{Numeric vector of trait values (e.g., mean_trait or sd_trait).}
}
\value{
Integer representing the optimal number of bins.
}
\description{
Calculates the optimal number of bins for histogramming a numeric vector
based on Scott's rule, which minimizes the integrated mean squared error.
}
