% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygons}
\alias{polygons}
\title{Species distribution polygons for 4 Carnivora species}
\format{
An \code{sf} object with the following columns:
\describe{
\item{TaxonName}{Species name (matching the \code{traits} table)}
\item{geometry}{Polygon geometry representing species distribution}
}
}
\source{
Donwload from the IUCN Red List webpage (IUCN, 2025).
}
\usage{
polygons
}
\description{
A spatial dataset of species range polygons matching the species in the \code{traits} dataset.
}
\keyword{datasets}
