% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits}
\alias{traits}
\title{Relative blade length trait data for Carnivora}
\format{
A data frame with the following columns:
\describe{
\item{TaxonName}{Species name (binomial)}
\item{RBL}{Relative blade length (unitless ratio)}
}
}
\source{
Siciliano-Martina et al. (2024). \emph{Ecology and Evolution}, 14(10), e70214.
}
\usage{
traits
}
\description{
A dataset of relative blade length (RBL) values for 10 species in the order Carnivora.
These species match those in the \code{polygons} dataset.
}
\keyword{datasets}
